third_party/WebKit/LayoutTests/http/tests/inspector-protocol/network-fetch-content-with-error-status-code-expected.txt

Test to make sure if an xhr is fetched with the response as a blob and cross origin devtools can get body.
Network Enabled
Evaled fetch command in page
Request will be sent
Request Method (should be OPTIONS): OPTIONS
Event URL has appropriate ending: true
Second Response Method (should be GET): GET
Got response received
requestId is the same as requestWillBeSent: true
Response Body: Some content
